Split off ring proofs module
